Quell Gcc 4.03 compiler warning
Returning -1 as an unsigned int warns as follows in GCC 4.03:

warning: converting negative value '-0x00000000000000001' to 'unsigned int'

git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@20021 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfsAPI.cxx
@@ -3383,7 +3383,7 @@ const Obstacle* findFirstNameInList ( const ObstacleList &list, const std::strin
 BZF_API unsigned int bz_findWorldObject ( const char *name )
 {
   if (!name)
-    return -1;
+    return (unsigned int)-1;
 
   std::string nameStr = name;
   for (int type = wallType; type < ObstacleTypeCount; type++) {
@@ -3399,7 +3399,7 @@ BZF_API unsigned int bz_findWorldObject ( const char *name )
     return obs->getGUID();
   }
 
-  return -1;
+  return (unsigned int)-1;
 }
 
 BZF_API bz_APIBaseWorldObject* bz_getWorldObjectByID ( unsigned int id )
